Make the list navigation buttons work
When a user interacts with the Previous or Next button, it will show the
next or previous page of the list, or will show an error if the user
tries to go out of bounds.

Note that the bot will reply to the message with the buttons (in other
words, its own message). It will also use the list of the user who is
interacting with the buttons, not the user who originally triggered the
list command.